+++ src/utils/helper/common.sh
@@ -83,17 +83,16 @@ _myzs:internal:load() {
   local args=("$@")
 
   if _myzs:internal:checker:file-exist "$filepath"; then
-
     _myzs:internal:log:debug "source ${filepath} with '${args[*]}'"
     source "${filepath}" "${args[@]}"
     exitcode=$?
     if [[ "$exitcode" != "0" ]]; then
       _myzs:internal:log:error "Cannot load ${filename} (${filepath}) because source return $exitcode"
       _myzs:internal:failed "$exitcode"
+    else
+      _myzs:internal:log:info "Loaded ${filename} (${filepath}) to the system"
+      _myzs:internal:completed
     fi
-
-    _myzs:internal:log:info "Loaded ${filename} (${filepath}) to the system"
-    _myzs:internal:completed
   else
     _myzs:internal:log:warn "Cannot load ${filename} (${filepath}) because file is missing"
     _myzs:internal:failed 4